changed from mgo aggregation pipeline to go loops

+++ types/tick.go
@@ -4,7 +4,7 @@ type Tick struct {
 	Block     int64 `json:"block" bson:"block"`
 	OpenTime  int64 `json:"opentime" bson:"opentime"`
 	CloseTime int64 `json:"closetime" bson:"closetime"`
-	Volume    int64 `json:"volume" bson:"volume"`
+	Volume    Int   `json:"volume" bson:"volume"`
 	Open      Int   `json:"open" bson:"open"`
 	Close     Int   `json:"close" bson:"close"`
 	High      Int   `json:"high" bson:"high"`
